+++ test/optimizer/level-0/optimizations-test.js
@@ -1,0 +1,14 @@
+var vows = require('vows');
+
+var optimizerContext = require('../../test-helper').optimizerContext;
+
+vows.describe('level 0')
+  .addBatch(
+    optimizerContext('optimizations', {
+      'are off': [
+        'a{color:#f00;font-weight:bold}p{color:#f00}',
+        'a{color:#f00;font-weight:bold}p{color:#f00}'
+      ]
+    }, { level: 0 })
+  )
+  .export(module);